+++ CHANGELOG.md
@@ -8,7 +8,9 @@ v0.7.0
 - [x] Added option for smooth scroll
 - [x] Preferences are now saved when updating the app
 - [x] Validators for smooth_scroll and preview_format
-- [ ] Display a "What's New" Popup when app is run after an update
+- [x] Display a "What's New" Popup when app is run after an update
+
+- - -
 
 v0.6.2
 ------